fixed pom, iniciar servidor desde linea de comandos

+++ src/main/java/com/satoshihans/practicalaboralsql/models/dto/TrabajaCreacionDTO.java
@@ -6,5 +6,5 @@ import lombok.*;
 @AllArgsConstructor
 public class TrabajaCreacionDTO {
     private Long id_especialista;
-    private Float importe;
+    private Double importe;
 }
